% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_summary}
\alias{compute_summary}
\title{Quantiles from generalized Pareto}
\usage{
compute_summary(data, thresh, p = 0.5, mlen = 1000L)
}
\arguments{
\item{data}{[numeric] vector of observations}

\item{thresh}{[numeric] vector of thresholds}

\item{p}{[numeric] scalar probability level for quantile}

\item{mlen}{[integer] scalar number of observations for the extrapolation.}
}
\description{
This function fits the generalized Pareto distribution
at each threshold and returns the estimated quantile of the \code{mlen}
\code{p}th quantile (default is median) as measure of risk.
The function is vectorized over \code{thresh}.
}
