% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_threshold.R
\name{find_threshold}
\alias{find_threshold}
\title{...}
\usage{
find_threshold(x, method = "rosin", x_eval = 101, x_pos = NULL, ...)
}
\arguments{
\item{x}{...}

\item{method}{...}

\item{x_eval}{...}

\item{...}{...}
}
\description{
...
}

