% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{ft}
\alias{ft}
\alias{dft}
\alias{qft}
\alias{pft}
\alias{rft}
\title{Folded student}
\usage{
dft(x, df, log = FALSE)

qft(p, df, lower.tail = TRUE, log.p = FALSE)

pft(q, df, lower.tail = TRUE, log.p = FALSE)

rft(n, df)
}
\arguments{
\item{x}{vector of quantiles.}

\item{df}{degrees of freedom}

\item{log}{logical; if \code{TRUE}, probabilities \code{p} are given as \eqn{\log(p)}.}

\item{p}{vector of probabilities.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{\Pr(X \le x)}, otherwise \eqn{\Pr(X > x)}.}

\item{log.p}{logical; if \code{TRUE}, probabilities \code{p} are given as \eqn{\log(p)}.}

\item{q}{vector of quantiles.}

\item{n}{number of observations.}
}
\description{
Density, distribution function, quantile function and random generation for the Student-t distribution with \code{df} degrees of freedom.
}
\section{Functions}{
\itemize{
\item \code{dft()}: density

\item \code{qft()}: density

\item \code{pft()}: density

\item \code{rft()}: density

}}
