% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickands.R
\name{gp.lmom}
\alias{gp.lmom}
\title{Estimation of generalized Pareto parameters via L-moments}
\usage{
gp.lmom(xdat, sorted = FALSE)
}
\arguments{
\item{xdat}{[numeric] vector of observations}

\item{sorted}{[logical] if \code{TRUE}, observations are sorted in increasing order}
}
\value{
a vector of length two with the scale and shape estimates
}
\description{
Given a sample of exceedances, compute the first four L-moments and
use L-skewness and L-scale to compute the scale and shape of the generalized Pareto distribution
}
