% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickands.R
\name{gp.pwm}
\alias{gp.pwm}
\title{Probability weighted moment estimator}
\usage{
gp.pwm(xdat, sorted = FALSE)
}
\arguments{
\item{xdat}{[numeric] vector of observations}

\item{sorted}{[logical] if \code{TRUE}, observations are sorted in increasing order}
}
\value{
a vector of length two with the scale and shape estimates
}
\description{
Given a vector of sample exceedances, returns a parameter estimates for the generalized Pareto distribution
based on the two first empirical probability weighted moments.
}
