% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKK.R
\name{mdps}
\alias{mdps}
\title{Minimum distance selection procedure}
\usage{
mdps(data)
}
\arguments{
\item{data}{vector of positive exceedances}
}
\value{
a list with components
\describe{
\item{\code{k0}}{order statistic corresponding to threshold (number of exceedances)}
\item{\code{tail.index}}{Hill's estimator of the tail index based on k0 exceedances}
\item{\code{threshold}}{numerical value of the threshold, the n-k0+1 order statistic of the original sample}
}
}
\description{
Minimum distance selection procedure
}
\references{
Clauset, A., Shalizi, C.R. and Newman, M.E.J. (2009). \emph{Power-Law Distributions in Empirical Data}. SIAM Review. Society for Industrial and Applied Mathematics, \bold{51}, 661-703, \url{https://doi.org/10.1137/070710111}
}
