% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickands.R
\name{pickands.diag}
\alias{pickands.diag}
\title{Pickands' order statistics selection method}
\usage{
pickands.diag(xdat, method = c("quartiles", "mle", "pwm", "lmom"))
}
\arguments{
\item{xdat}{[numeric] vector of observations}

\item{method}{[string] estimation method, either the quartiles of Pickands (1975), maximum likelihood, probability weighted moments or L-moments}
}
\value{
a vector with elements \code{k}, the number of order statistics to keep, and \code{thresh}, the numerical value of the latter
}
\description{
Restricting to the largest fourth of the data, consider
the minimum Kolmogorov-Smirnov statistic, i.e., the maximum
absolute difference between the estimated generalized Pareto
and the empirical distribution of exceedances.
}
\references{
James Pickands III (1975). \emph{Statistical inference using extreme order statistics}, Annals of Statistics, 3(\bold{1}) 119-131, \url{https://doi.org/10.1214/aos/1176343003}
}
