% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.threshold_list.R
\name{plot.threshold_list}
\alias{plot.threshold_list}
\title{...}
\usage{
\method{plot}{threshold_list}(x, ths_add = NULL, names_add = NULL,
  add2hist = FALSE, ylim = NULL, ...)
}
\arguments{
\item{x}{...}

\item{ths_add}{...}

\item{names_add}{...}

\item{add2hist}{...}

\item{ylim}{...}

\item{...}{...}
}
\description{
...
}

