% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{prjbmixt}
\alias{prjbmixt}
\title{Gumbel-generalized Pareto mixture with contamination}
\usage{
prjbmixt(q, kappa, beta, sigma, xi, u, eps, lower.tail = TRUE)
}
\arguments{
\item{kappa}{shape of Weibull}

\item{beta}{scale of Weibull}

\item{sigma}{scale of generalized Pareto}

\item{xi}{shape of generalized Pareto}

\item{u}{threshold above which mixture}

\item{eps}{amount of overlap}

\item{lower.tail}{logical; if \code{TRUE}, return distribution function and survival function otherwise}
}
\description{
The model follows a Weibull distribution below \code{u} and a generalized Pareto above \code{u}, but blends continuously on the interval [\code{u}, \code{u}+\code{eps}]
}
\examples{
q <- seq(0, 16, length.out = 200)
plot(x = q,
     y = prjbmixt(q = q,
          kappa = 0.7,
          beta = 2,
          sigma = 2,
          xi = (1-0.7)/0.7*((9/2)^(-0.7)),
          u = 9,
          eps = 1),
    xlab = "quantile",
    ylab = "distribution function",
    main = "Weibull-GP mixture",
    type = "l")
}
\references{
Roth, Jongbloef, Buishand (2016), \emph{Threshold selection for regional peaks-over-threshold data}, Journal of Applied Statistics
}
\keyword{internal}
