% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{qft}
\alias{qft}
\title{Quantile function of folded Student-t}
\usage{
qft(p, df = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of probabilities}

\item{df}{degrees of freedom}

\item{lower.tail}{logical; if \code{TRUE}, returns the lower tail (distribution function)}

\item{log.p}{logical; if \code{TRUE}, argument \code{p} contains log-probabilities.}
}
\description{
Quantile function of folded Student-t
}
\keyword{internal}
