% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rft}
\alias{rft}
\title{Random generation for folded Student-t}
\usage{
rft(n, df)
}
\arguments{
\item{n}{sample size}
}
\description{
Generate random sample from half-Student above zero via folding
}
\keyword{internal}
