% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{risk_measures}
\alias{risk_measures}
\title{Risk measures for selected parametric models}
\usage{
risk_measures(
  model,
  p = 0.999,
  risk = c("quantile", "maxquant", "penultimate"),
  mlen = 1L
)
}
\arguments{
\item{model}{[list] a list with argument \code{family} and additional arguments within list \code{args}}

\item{p}{[double] probability level for quantile and expectiles}

\item{risk}{[string] choice of risk function, either `quantile` or `maxquant` for `penultimate`}

\item{mlen}{[integer] length of block maximum}
}
\description{
Given a parametric distribution, compute selected risk
measures. These are used to benchmark the threshold methods
by comparing the bias, variance, mean absolute error,
root mean squared error relative to an oracle that picks the
threshold with the best performance, or the correct model if the tail
is exactly generalized Pareto above a threshold
}
\examples{
# Compute return level of 1000 obs
risk_measures(
 model = list(family = "weibull",
              args = list(shape = 1.25)),
 p = 0.366,
 m = 1000L,
 risk = "maxquant")
risk_measures(
 model = list(family = "weibull",
              args = list(shape = 1.25)),
 p = 1-1/1000,
 risk = "quantile")
risk_measures(
 model = list(family = "weibull",
              args = list(shape = 1.25)),
 p = 0.366,
 mlen = 1000L,
 risk = "penultimate")
}
