% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKK.R
\name{samsee}
\alias{samsee}
\title{Threshold selection via SAMSEE}
\usage{
samsee(data)
}
\arguments{
\item{data}{vector of positive exceedances}
}
\value{
a list with elements
\describe{
\item{\code{k}}{optimal number of exceedances}
\item{\code{gamma_hill}}{Hill's estimator of the tail index}
\item{\code{gamma_v}}{de Vries estimator of the tail index}
\item{\code{gamma_gj}}{generalized jackknife estimator of the tail index}
}
}
\description{
Smooth asymptotic mean squared error estimator
of Schneider et al. (2021) for threshold selection.
The implementation uses a second-order regular variation index of -1.
}
\references{
Schneider, L.F., Krajina, A. & Krivobokova, T. (2021). \emph{Threshold selection in univariate extreme value analysis}, Extremes, \bold{24}, 881–913 \url{https://doi.org/10.1007/s10687-021-00405-7}
}
