% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKK.R
\name{sihs}
\alias{sihs}
\title{Smooth inverse Hill threshold selection}
\usage{
sihs(data, maxk = 500L)
}
\arguments{
\item{data}{vector of positive exceedances}

\item{maxk}{maximum number of exceedances for the method, default to 500.}
}
\value{
a list with components
\describe{
\item{\code{k0}}{order statistic corresponding to threshold (number of exceedances)}
\item{\code{tail.index}}{Hill's estimator of the tail index based on k0 exceedances}
\item{\code{threshold}}{numerical value of the threshold, the n-k0+1 order statistic of the original sample}
}
}
\description{
Threshold selection via smooth inverse Hill statistic using \code{eBsc} nonparametric smoothing. As the latter is computationally intensive, it is recommended to limit the number of order statistics considered
}
\references{
Schneider, L.F., Krajina, A. & Krivobokova, T. (2021). \emph{Threshold selection in univariate extreme value analysis}, Extremes, \bold{24}, 881–913 \url{https://doi.org/10.1007/s10687-021-00405-7}
}
