% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surprise.R
\name{surprise.diag}
\alias{surprise.diag}
\title{Bayesian measure of surprise}
\usage{
surprise.diag(
  dat,
  u,
  type = c("posterior", "partial"),
  prior = "mdi",
  B = 1000L,
  nrep = 100,
  stat = c("reciplik", "quantile", "rlargest"),
  os = 1,
  plot = TRUE
)
}
\arguments{
\item{dat}{[vector] data}

\item{u}{[vector] threshold}

\item{type}{[string] type of likelihood, either \code{partial} or \code{posterior}. See the article for references}

\item{prior}{[string] name of prior for the \code{revdbayes} package (default to maximal data information)}

\item{B}{[integer] number of posterior replications (default to 10 000)}

\item{nrep}{[integer] number of replications for the p-value calculation}

\item{stat}{[string] the statistic, either reciprocal likelihood (\code{reciplik}) or jth empirical quantile(\code{quantile})}

\item{os}{[integer] order statistic for the \code{quantile} statistic. Default to maximum (\code{os=1})}

\item{plot}{[logical] produce a plot of the Bayesian p-value against threshold}
}
\value{
an invisible list with elements
\describe{
\item{\code{threshold}}{\code{k} vector of candidate thresholds}
\item{\code{pval}}{\code{k} by \code{nrep} matrix of p-values}
\item{\code{mean_pval}}{average p-value for each threshold},
\item{\code{stat}}{string indicating the test statistic}
\item{\code{type}}{string indicating whether the partial or full posterior are used}
\item{\code{nrep}}{integer number of replication}
}
}
\description{
Threshold stability plots proposed by Lee, Fan and Sisson (2015) using measures of goodness-of-fit for threshold selection.
Values around 1/2 imply that the generalized Pareto model is a good approximation to the threshold exceedances.
}
\details{
The statistics are computed in three steps for each threshold:
\itemize{
\item Sample from the posterior distribution of the scale and shape parameters
\item For each parameter pair drawn of the posterior, simulate a sample from the posterior predictive distribution with as many observations as the number of exceedances above the threshold
\item Evaluate the statistic for both the exceedances and the simulated sample
\item Repeat a large number of time the previous steps.
\item Compute Bayesian p-value as proportion of times the value exceeds that for observed data
}
}
