% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/threeGaussians.R
\name{threeGaussians}
\alias{threeGaussians}
\title{...}
\usage{
threeGaussians(n = c(1000, 1000, 4000), m.pos = 2, m.neg = c(-10, -4),
  sd.pos = 1, sd.neg = c(1, 2), seed = NULL)
}
\arguments{
\item{n}{...}

\item{m.pos}{...}

\item{m.neg}{...}

\item{sd.pos}{...}

\item{sd.neg}{...}

\item{seed}{...}
}
\description{
...
}

