% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/threshold_entropy.R
\name{threshold_entropy}
\alias{threshold_entropy}
\title{Entropy-based thresholding}
\usage{
threshold_entropy(x, range_constrain = NULL, ...)
}
\arguments{
\item{x}{vector of (pixel) gray levels to be thresholded}

\item{range_constrain}{constrain the range from which to select the threshold}

\item{...}{arguemnts passed to hist, e.g. breaks}
}
\description{
Entropy based thresholding based on Kapur et al. (1985).
}

