% Generated by roxygen2 (4.0.1): do not edit by hand
\name{threshold_valleys}
\alias{threshold_valleys}
\title{Thresholding based on peak and valley analysis.}
\usage{
threshold_valleys(x, x_eval = 101, ...)
}
\arguments{
\item{x}{vector of (pixel) gray levels to be thresholded}

\item{x_eval}{...}

\item{...}{arguemnts passed to peaks_and_valleys}
}
\description{
A smooth density model is derived for the data with \code{\link[logspline]{logspline}}. Based on the density peaks and valleys are detected and the final threshold is defined as the valley  (if available) with the highest \code{x} value.
}

