% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoGaussians.R
\name{twoGaussians}
\alias{twoGaussians}
\title{...}
\usage{
twoGaussians(n = c(1000, 5000), m.pos = 2, m.neg = -2, sd.pos = 1,
  sd.neg = 1, seed = NULL)
}
\arguments{
\item{n}{...}

\item{m.pos}{...}

\item{m.neg}{...}

\item{sd.pos}{...}

\item{sd.neg}{...}

\item{seed}{...}
}
\description{
...
}

