% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.threshold.R
\name{plot.threshold}
\alias{plot.threshold}
\title{...}
\usage{
\method{plot}{threshold}(x, add2hist = FALSE, ylim = NULL, ...)
}
\arguments{
\item{x}{...}

\item{add2hist}{...}

\item{ylim}{...}

\item{...}{...}
}
\description{
...
}

