% Generated by roxygen2 (4.0.1): do not edit by hand
\name{threshold_nmm}
\alias{threshold_nmm}
\title{Thresholding based on normal mixture modelling.}
\usage{
threshold_nmm(x, x_eval = 101, ...)
}
\arguments{
\item{x}{vector of (pixel) gray levels to be thresholded, or an object of class densityMclust Mclust}

\item{x_eval}{points which}

\item{...}{arguemnts passed to mclust}
}
\description{
It is assumed that the data can be descrived by two Gaussians.
The parameters of the mixture components are estimated with \code{\link[mclust]{Mclust}}.
}
\references{
...
}

