% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartymetric.R
\name{vmetric.diag}
\alias{vmetric.diag}
\title{Metric-based threshold selection}
\usage{
vmetric.diag(
  xdat,
  thresh,
  B = 199L,
  type = c("exp", "qq", "pp"),
  dist = c("l1", "l2"),
  paruq = FALSE,
  neval = 1000L
)
}
\arguments{
\item{xdat}{[numeric] vector of observations}

\item{thresh}{[numeric] vector of thresholds}

\item{B}{[integer] number of bootstrap replications}

\item{type}{[string] type of graph, either \code{pp} for probability-probability plots, \code{qq} for quantile-quantile plots or \code{exp} for unit exponential Q-Q plots}

\item{dist}{[string] the distance used, either absolute distance (\code{l1}) or Euclidean distance (\code{l2})}

\item{paruq}{[logical] if \code{TRUE}, generate bootstrap samples accounting for the sampling distribution of parameters}

\item{neval}{[integer] number of points at which to estimate the metric. Default to 1000}
}
\description{
Poor man's adaptation of Varty et al. for the
independent and identically distributed case
with no rounding.
}
\details{
The algorithm proceeds by first computing the maximum
likelihood algorithm and then simulating datasets from
replication with parameters drawn from a bivariate normal
approximation to the maximum likelihood estimator distribution.

For each bootstrap sample, we refit the
 model and convert the quantiles to
exponential or uniform variates.
The mean absolute or mean squared distance
is calculated on these. The threshold
returned is the one with the lowest value
of the metric.
}
